fix(/articles/): update publication date of article
Since the new codebase which uses src/root/, the article index.md
file has been moved to this directory; that means that the
creation date of the file is for Git the date of the commit which
moved it.

The only solution to this is to change the article publication
date.